# educa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "educa API",
        "description": "educa API description",
        "contact": {
            "email": "kontakt@schule-plus.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/administration/widgets/contacts": {
            "get": {
                "tags": [
                    "administration",
                    "v1",
                    "widgets"
                ],
                "description": "",
                "operationId": "aeff2f8411380cfb1decbe13d0218b2c",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "jwt token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "school",
                        "in": "query",
                        "description": "id of the school",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns data for a contact graph"
                    }
                }
            }
        },
...
        "/api/v1/cloud_user/all": {
            "get": {
                "tags": [
                    "v1",
                    "cloud_user"
                ],
                "description": "",
                "operationId": "5b68041e86cc2b4a5e8fd9030cd44f67",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all cloud users to build a cache "
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "v1",
                    "me"
                ],
                "description": "",
                "operationId": "05dddd820647164cc4250d48e75a1ec7",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information about the current user according to the token including information about the avaiable apps and group memebership"
                    }
                }
            }
        },
...
        "/api/v1/groups/apps/all": {
            "get": {
                "tags": [
                    "v1",
                    "groups",
                    "group_apps"
                ],
                "description": "",
                "operationId": "b0cbafbdf128701a67c7a0a62a2c82c3",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all apps that exist"
                    }
                }
            }
        },
...
    }
}